+++ lib/zaikio/hub/specialist.rb
@@ -7,7 +7,7 @@ module Zaikio
       include_root_in_json :specialist
 
       # Attributes
-      attributes :name, :updated_at, :created_at, :site_id, :kind, :owner_id
+      attributes :name, :updated_at, :created_at, :site_id, :kind, :owner_id, :capabilities
     end
   end
 end